aider: Rename `get_tasks_from_disk` to `_get_tasks_from_disk` and update the call to this function in the `run` method.
# Aider chat conversation:

USER: make the get_tasks_from_disk function private

ASSISTANT: To make the `get_tasks_from_disk` function private, we need to rename it to `_get_tasks_from_disk`. This is a convention in Python to indicate that a method or attribute is intended to be private.

Here is the *edit block* for this change:

<source>python
metagpt/actions/write_project_approval.py
<<<<<<< HEAD
    def get_tasks_from_disk(self):
=======
    def _get_tasks_from_disk(self):
>>>>>>> updated
</source>

Also, we need to update the call to this function in the `run` method:

<source>python
metagpt/actions/write_project_approval.py
<<<<<<< HEAD
            output = self.get_tasks_from_disk()
=======
            output = self._get_tasks_from_disk()
>>>>>>> updated
</source>

Please confirm if this is the change you want to make.